Added `panic_info` feature
