+++ kernel/arch/silicium-x86_64/src/lib.rs
@@ -1,5 +1,5 @@
 #![cfg_attr(not(test), no_std)]
-#![feature(panic_info_message)]
+#![cfg_attr(feature = "panic_info", feature(panic_info_message))]
 #![feature(negative_impls)]
 #![feature(const_pin)]
 
@@ -12,6 +12,7 @@ pub mod gdt;
 pub mod idt;
 pub mod io;
 pub mod irq;
+pub mod lang;
 pub mod msr;
 pub mod opcode;
 pub mod paging;
@@ -61,33 +62,4 @@ pub unsafe fn setup() {
 
     smp::setup();
     paging::load_kernel_pml4();
-}
-
-#[cfg(not(test))]
-#[panic_handler]
-pub fn panic(info: &core::panic::PanicInfo) -> ! {
-    use crate::apic::local::{IpiDestination, IpiPriority};
-
-    irq::disable();
-    if smp::ap_booted() {
-        // SAFETY: This is safe because we have ensured that the APs are booted, meaning
-        // that they can safely receive IPIs without triple faulting.
-        unsafe {
-            apic::local::send_ipi(IpiDestination::AllExcludingSelf, IpiPriority::Nmi, 0x00);
-        }
-    }
-
-    log::error!("The kernel has encountered a fatal error that it cannot recover from");
-    log::error!("The kernel must stop to prevent further damage");
-
-    if let Some(message) = info.message() {
-        if let Some(location) = info.location() {
-            log::error!("CPU {}: {} at {}", cpu::id(), message, location);
-        } else {
-            log::error!("{}", message);
-        }
-    }
-
-    // Halt the CPU
-    cpu::halt();
 }
